refactor(extractor): moving source extraction logic outside classes
