--- src/scripts/content/extractor/sources/ExtractorSources.js
@@ -1,5 +1,0 @@
-export default class ExtractorSources {
-  async extractSources(content, format) {
-    throw new Error("Not implemented");
-  }
-}
